@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { CssBaseline, Grid, ThemeProvider } from '@mui/material';
-import { createTheme } from '@mui/material/styles';
+import { CssBaseline, Grid } from '@mui/material';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
 
 import { selectCenterPosition, selectZoom, selectUITheme} from './store/uiparamSlice';
 
@@ -12,13 +12,13 @@ import Menubar from './Menubar';
 /* Style info to use for theming */
 const darkTheme = createTheme({
   palette: {
-    type: 'dark',
+    mode: 'dark',
   }
 });
 
 const lightTheme = createTheme({
   palette: {
-    type: 'light',
+    mode: 'light',
   }
 });
 
@@ -35,18 +35,18 @@ export default function App(props) {
   const UITheme = useSelector(selectUITheme);
 
   return (
-
-    <ThemeProvider theme={UITheme === 'dark' ? darkTheme : lightTheme}>
-      <CssBaseline />
-      <Grid container direction='column' spacing={1} style={{ height: '100%' }}>
-        <Grid item style={{ width: '100%' }}>
-            <Menubar/>
-        </Grid>
-        <Grid item style={{ flexGrow: 1, width: '100%' }}>
-          <MapViewer initposition={position} initZoom={zoom} />
-        </Grid>
-      </Grid>
-    </ThemeProvider>
-
+      <React.StrictMode>
+        <ThemeProvider theme={UITheme === 'dark' ? darkTheme : lightTheme}>
+          <CssBaseline />
+          <Grid container direction='column' spacing={1} style={{ height: '100%' }}>
+            <Grid item style={{ width: '100%' }}>
+                <Menubar/>
+            </Grid>
+            <Grid item style={{ flexGrow: 1, width: '100%' }}>
+              <MapViewer initposition={position} initZoom={zoom} />
+            </Grid>
+          </Grid>
+        </ThemeProvider>
+      </React.StrictMode>
   );
 }
